reorganization of client folder
