+++ client/web/package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^7.1.2",
     "deck.gl": "^8.2.8",
     "mapbox-gl": "^1.3.1",
+    "node-sass": "^4.14.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-map-gl": "^5.2.8",
