+++ client/web/src/App.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Header from './components/main/Header';
+import Routes from './Routes';
+
+const App = () => {
+  return (
+    <>
+      <Header />
+      <Routes />
+    </>
+  );
+}
+
+export default App;
