+++ client/web/src/components/main/Header.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const Header = () => {
+  return (
+    <>
+    </>
+  );
+}
+
+export default Header;
