+++ client/web/src/components/map/SearchResult.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import "./SearchResult.css";
 
 export default function SearchResult(props) {
   return (
