+++ client/web/src/styles/_global.scss
@@ -26,7 +26,7 @@ h1 {
 
 @font-face {
   font-family: adobe-caslon-pro;
-  src: url(../src/assets/ACaslonPro-Bold.otf);
+  src: url(../assets/ACaslonPro-Bold.otf);
 }
 
 h2 {
