+++ client/web/src/styles/map/SearchResult.scss
