+++ client/web/src/styles/map/SidePanelList.scss
