+++ client/web/src/styles/map/SidePanelSelection.scss
