+++ client/web/src/styles/styles.scss
@@ -1,0 +1,6 @@
+@charset "utf-8";
+
+@import './global';
+
+@import './main/about';
+@import './main/home';
